NKS: bug fixes.

+++ brihaspati/trunk/BGAS/config/sqlscripts/createtable_bgasAccData-20131219.sql
@@ -2,12 +2,12 @@ use login;
 
 CREATE TABLE IF NOT EXISTS bgasAccData
 (
-                        id INTEGER (11)  NOT NULL AUTO_INCREMENT,
+                        id int(11)  NOT NULL AUTO_INCREMENT,
                         organization VARCHAR (255) NOT NULL,
                         unit varchar (255) NOT NULL,
                         databasename VARCHAR (255) NOT NULL,
                         year VARCHAR (4) NOT NULL,
                         PRIMARY KEY(id),
                         UNIQUE(databasename)
-)ENGINE=InnoDB DEFAULT CHARSET=latin1;
+) ENGINE=InnoDB DEFAULT CHARSET=latin1;
